Move Pulumi-related code into its own folder
This makes space for testing another IaC tool or adding anything
else we need without intermingling with the Pulumi project.
